Paing for an order
Now paying for an order generates a relational table for the order with details and has an id. That id is a column in orderitems table where the specific items are shown. Next we could make an interface/UI for seeing orders, so that a firm could manage orders from a UI.